+++ springboot-bootcamp/l-community/src/main/java/com/lawencon/community/model/PriceListEvent.java
@@ -1,5 +1,7 @@
 package com.lawencon.community.model;
 
+import java.math.BigInteger;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
@@ -23,7 +25,7 @@ public class PriceListEvent extends BaseEntity {
 	private String priceCode;
 	
 	@Column(name="price_nominal")
-	private Integer priceNominal;
+	private BigInteger priceNominal;
 	
 	public String getPriceName() {
 		return priceName;
@@ -37,10 +39,10 @@ public class PriceListEvent extends BaseEntity {
 	public void setPriceCode(String priceCode) {
 		this.priceCode = priceCode;
 	}
-	public Integer getPriceNominal() {
+	public BigInteger getPriceNominal() {
 		return priceNominal;
 	}
-	public void setPriceNominal(Integer priceNominal) {
+	public void setPriceNominal(BigInteger priceNominal) {
 		this.priceNominal = priceNominal;
 	}
